cleared warning
git-svn-id: svn://svn.cc65.org/cc65/trunk@1386 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ src/grc/grc.c
@@ -50,7 +50,7 @@ unsigned char buffer[BLOODY_BIG_BUFFER];
 unsigned char vlirtabt[127];
 unsigned char vlirtabs[127];
 int i,j;
-size_t bytes;
+int bytes;
 int blocks,rest;
 
 	i=2;
